memindahkan properti spouse dan children
memindahkan properti spouse dan children ke class PersonalInfo
