+++ src/main/java/Employee.java
@@ -7,22 +7,12 @@ public class Employee {
     private PersonalInfo personalInfo;
     private EmploymentDetails employmentDetails;
 
-    private Spouse spouse;
-    private List<Child> children;
+
 
     public Employee(String employeeId,PersonalInfo personalInfo, EmploymentDetails employmentDetails) {
         this.employeeId = employeeId;
         this.personalInfo = personalInfo;
         this.employmentDetails = employmentDetails;
-        this.children = new LinkedList<Child>();
-    }
-
-    public void setSpouse(Spouse spouse) {
-        this.spouse = spouse;
-    }
-
-    public void addChild(Child child) {
-        this.children.add(child);
     }
 
     public int getAnnualIncomeTax() {
