+++ src/main/java/PersonalInfo.java
@@ -1,9 +1,14 @@
+import java.util.LinkedList;
+import java.util.List;
+
 public class PersonalInfo {
     private String firstName;
     private String lastName;
     private String idNumber;
     private String address;
     private boolean gender; //true = Laki-laki, false = Perempuan
+    private Spouse spouse;
+    private List<Child> children;
 
     public PersonalInfo(String firstName, String lastName, String idNumber, String address, boolean gender) {
         this.firstName = firstName;
@@ -11,6 +16,7 @@ public class PersonalInfo {
         this.idNumber = idNumber;
         this.address = address;
         this.gender = gender;
+        this.children = new LinkedList<Child>();
     }
 
     public String getFirstName() {
@@ -31,5 +37,13 @@ public class PersonalInfo {
 
     public boolean isGender() {
         return gender;
+    }
+
+    public void setSpouse(Spouse spouse) {
+        this.spouse = spouse;
+    }
+
+    public void addChild(Child child) {
+        this.children.add(child);
     }
 }
